Move TensorLayout to dali/core.
